reduce mem footprint

+++ src/atom.rs
@@ -1,11 +1,11 @@
 #[derive(Debug)]
 pub struct Nucleus {
-    pub protons:  i64,
-    pub neutrons: i64,
+    pub protons:  i8,
+    pub neutrons: i8,
 }
 
 #[derive(Debug)]
 pub struct Atom {
-    pub electrons: i64,
+    pub electrons: i8,
     pub nucleus: Nucleus,
 }
